Add short deck ranker
The equivalence classes are mechanically and manually adjusted from the
full deck. This presumes the current style of properly adjusting
flushes above full house.. but leaving straights ahead of
three-of-a-kind.

This has really driven compile times through the roof, but that's the
price we pay, I guess.

Up-coming, then, a rethink of the evaluation atom as an options list,
including which deck to use and how to pick from community and personal
cards.